requirements.txt
Pillow==10.4.0
PyAutoGUI==0.9.54
pytesseract==0.3.10
torch==2.4.0+cu124
transformers==4.43.3
